+++ packages/webawesome/docs/docs/resources/changelog.md
@@ -35,9 +35,9 @@ Web Awesome follows <a href="https://semver.org/" class="appearance-plain">Seman
 
 - Fixed type resolution issues with `pro` components like `<wa-file-input>`, `<wa-combobox>`, etc. [pr:2577]
 - Fixed an issue where some fonts wouldn't load in themes that import multiple fonts [pr:2582]
+- Fixed `<wa-page>` rendering an empty footer in its mobile navigation drawer when no `navigation-footer` is slotted, which left dead space below the sidebar
 
 :::
-
 
 ## 3.10.0